+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next version
 * Add `::all()` to the kind type to iterate over all kind variants
 * Generate customizable implementation of `Display` trait
+* Generate implementation of `FromStr` trait
 
 ## v0.0.3 - 2023-08-05
 * Make generated `kind()` function public